+++ push_swap/libft/ft_count_num_long_long.c
@@ -6,16 +6,16 @@
 /*   By: medesmon <medesmon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/20 04:32:21 by medesmon          #+#    #+#             */
-/*   Updated: 2019/09/20 04:41:53 by medesmon         ###   ########.fr       */
+/*   Updated: 2019/10/17 16:28:42 by medesmon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_count_num_long_long(unsigned long long n)
+int	ft_count_num_long_long(int n)
 {
-	long				i;
-	unsigned long long	neg;
+	int	i;
+	int	neg;
 
 	i = 0;
 	neg = n;
